created recipeMap util function for storing weights for the recipes
